Remove faulty local_networks defintion in rspam options #6931
The local_networks point to a non-existing map file, which isn't a problem in itself, but if both local_addrs and local_networks is set, rspamd ignores the local_addrs configuration. This results in all IPs (e.g. private IP ranges and localhosts by default) don't disable a whole bunch of checks on outbound mails.